refactor(types): move the interfaces to other files to easy reusable
