+++ src/types/github.ts
@@ -1,0 +1,13 @@
+export interface GithubOauthApp {
+  name: string;
+  clientId?: string;
+  redirectUrl: string;
+  clientSecretId?: string;
+  scope?: string;
+}
+
+export interface GithubAccessData {
+  accessToken: string;
+  scope: string;
+  tokenType: string;
+}
