+++ src/types/org.ts
@@ -1,0 +1,9 @@
+export interface Org {
+  name: string;
+  description: string;
+  avatarUrl: string;
+}
+
+export interface OrgsList {
+  [index: number]: Org;
+}
