+++ src/types/user.ts
@@ -1,0 +1,6 @@
+export interface User {
+  name: string;
+  login: string;
+  avatarUrl: string;
+  homepage: string;
+}
